Extract DeviceListProvider abstraction for async device enumeration
Introduced by: The isDeviceAvailable() pre-check in connectSource() runs
adb devices -l / pymobiledevice3 usbmux list synchronously on the UI
thread, blocking for up to 8 seconds (3s start + 5s finish timeout).
Both ADB and iOS have the identical blocking pattern.

## Problem

AdbLogcatSource::connectSource() calls transport_->isDeviceAvailable()
before connectTransport().  This spawns a subprocess synchronously:
  - ADB: `adb devices -l` (3s waitForStarted + 5s waitForFinished)
  - iOS: `pymobiledevice3 usbmux list` (same timeouts)
The UI thread freezes for the full duration on every connect/reconnect
attempt.

## Root cause

The blocking subprocess call lives directly in the transport classes
(AdbProcessTransport / IosLogProcessTransport) as a virtual override
of LiveSourceTransport::isDeviceAvailable().  There is no abstraction
to isolate the subprocess from the caller.

## How to fix

Create a DeviceListProvider abstraction:

- DeviceListProviderBase<T>: template base (no Q_OBJECT — MOC does not
  support templates) providing listDevices(), listDevicesAsync() (via
  QtConcurrent::run), and isDeviceAvailable().
- AdbDeviceListProvider: concrete provider for `adb devices -l`.
- IosDeviceListProvider: concrete provider for `pymobiledevice3 usbmux
  list`.

The blocking subprocess logic is moved from the transport classes into
the providers.  Transport classes own a provider instance and delegate
to it.  The static listDevices() methods are kept as thin delegates for
backward compatibility with dialog callers.

The isDeviceAvailable() pre-check is removed from connectSource()
because connectTransport() already handles device-not-found errors
gracefully.  The pre-check was redundant and blocked the UI thread.

## Files changed

New files:
- devicelistprovider.h — template base with sync/async device listing
- adbdeviceinfo.h — AdbDeviceInfo struct (extracted from transport)
- adbdevicelistprovider.h/.cpp — ADB provider (moved from transport)
- iosdevicelistprovider.h/.cpp — iOS provider (moved from transport)

Modified:
- adbprocesstransport.h/.cpp — owns AdbDeviceListProvider, removed
  isDeviceAvailable(), delegates listDevices() to provider
- ioslogprocesstransport.h/.cpp — owns IosDeviceListProvider, same
- livesourcetransport.h — removed virtual isDeviceAvailable()
- adblogcatsource.cpp — removed isDeviceAvailable() pre-check from
  connectSource()
- adb_ui_transport_test.cpp — 5 new tests for provider abstraction

## Tests added

- AdbDeviceListProvider returns same results as static listDevices
- AdbDeviceListProvider isDeviceAvailable returns true on subprocess error
- AdbDeviceListProvider listDevicesAsync returns a valid future
- IosDeviceListProvider isDeviceAvailable returns true on subprocess error
- IosDeviceListProvider listDevicesAsync returns a valid future

How to verify: `cmake --build build_root --target klogg_tests && \
./build_root/output/klogg_tests -platform offscreen` — 264 tests pass
(5487 assertions).
